refactor reward section
