+++ packages/atlas/src/views/global/YppLandingView/YppLandingView.styles.ts
@@ -5,7 +5,6 @@ import topRightPattern from '@/assets/images/ypp-background-pattern-2.svg'
 import bottomLeftPattern from '@/assets/images/ypp-background-pattern.svg'
 import { GridItem, LayoutGrid } from '@/components/LayoutGrid'
 import { LimitedWidthContainer } from '@/components/LimitedWidthContainer'
-import { Button } from '@/components/_buttons/Button'
 import { cVar, media, sizes } from '@/styles'
 
 export const imageShadow = css`
@@ -66,29 +65,6 @@ export const BackgroundContainer = styled.div<BackgroundContainerProps>`
     padding: ${sizes(24)} var(--size-global-horizontal-padding);
     ${backgroundPattern};
   }
-`
-export const BenefitsCardButton = styled(Button)`
-  border-radius: 999px;
-`
-
-export const BenefitsCardsButtonsGroup = styled.div`
-  text-align: center;
-  overflow-x: auto;
-  white-space: nowrap;
-  margin: ${sizes(16)} 0 ${sizes(8)} 0;
-
-  ::-webkit-scrollbar {
-    display: none;
-  }
-
-  ${BenefitsCardButton}:not(:last-child) {
-    margin-right: ${sizes(2)};
-  }
-`
-
-export const BenefitsCardsContainerGridItem = styled(GridItem)`
-  display: grid;
-  gap: ${sizes(4)};
 `
 
 export const CardsLimitedWidtContainer = styled(LimitedWidthContainer)`
